% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-layers.R
\name{Activation}
\alias{Activation}
\title{Activation Layer}
\usage{
Activation(activation = activation_identity, name = "")
}
\arguments{
\item{activation}{(defaults to `activation_identity`) – function to apply
at the end, e.g. `relu`}

\item{name}{(str, defaults to '') – the name of the function instance in
the network}
}
\value{
A function that accepts one argument and applies the operation to it
}
\description{
Layer factory function to create an activation layer. Activation functions
can be used directly in CNTK, so there is no difference between `y = relu(x)`
and `y = Activation(relu)(x)`. This layer is useful if one wants to configure
the activation function with default <- options, or when its invocation
should be named.
}
\examples{
```
model <- dense(500) \%>\% activation(activation_relu)()
# is the same as
model <- dense(500) \%>\% actvation_relu
# and also the same as
model <- dense(500, activation=activation_relu)
```
}
