% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-layers.R
\name{activation_identity}
\alias{activation_identity}
\title{Identity Function For CNTK Operations}
\usage{
activation_identity(keep)
}
\arguments{
\item{keep}{}
}
\description{
Identity Function For CNTK Operations
}
