% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{arraymixin_as_array}
\alias{arraymixin_as_array}
\title{ArrayMixin As Array}
\usage{
arraymixin_as_array(arraymixin)
}
\arguments{
\item{arraymixin}{- ArrayMixin instance}
}
\description{
ArrayMixin As Array
}
