% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{arrayview_from_data}
\alias{arrayview_from_data}
\title{ArrayView From Data}
\usage{
arrayview_from_data(data, device = NULL, read_only = FALSE,
  borrow = FALSE)
}
\arguments{
\item{data}{matrix to be converted}

\item{device}{instance of DeviceDescriptor}

\item{read_only}{(bool) whether the data can be modified}

\item{borrow}{(bool) whether nd_array memory can be borrowed internally to
speed up data creation}
}
\description{
Create an NDArrayView instance from a sparse matrix in CSR format.
}
