% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{arrayview_from_dense}
\alias{arrayview_from_dense}
\title{ArrayView From Dense Data}
\usage{
arrayview_from_dense(np_array, device = NULL, read_only = FALSE,
  borrow = FALSE)
}
\arguments{
\item{device}{instance of DeviceDescriptor}

\item{read_only}{(bool) whether the data can be modified}

\item{borrow}{(bool) whether nd_array memory can be borrowed internally to
speed up data creation}

\item{matrix}{matrix}
}
\description{
Create an NDArrayView instance from a matrix.
}
