% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{as_block}
\alias{as_block}
\title{As Block}
\usage{
as_block(composite, block_arguments_map, block_op_name,
  block_instance_name = "")
}
\arguments{
\item{block_instance_name}{}
}
\description{
Create a new block Function instance which just encapsulates the specified
composite Function to create a new Function that appears to be a primitive.
All the arguments of the composite being encapsulated must be Placeholder
variables. The purpose of block Functions is to enable creation of
hierarchical Function graphs where details of implementing certain building
block operations can be encapsulated away such that the actual structure of
the block’s implementation is not inlined into the parent graph where the
block is used, and instead the block just appears as an opaque primitive.
Users still have the ability to peek at the underlying Function graph that
implements the actual block Function.
}
