% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{as_composite}
\alias{as_composite}
\title{As Composite}
\usage{
as_composite(root_function, name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Creates a composite Function that has the specified root_function as its
root. The composite denotes a higher-level Function encapsulating the entire
graph of Functions underlying the specified rootFunction.
}
