% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{asvalue}
\alias{asvalue}
\title{Value Form Matrix}
\usage{
asvalue(variable, data_array)
}
\arguments{
\item{variable}{variable}

\item{data_array}{matrix of data}
}
\description{
Converts a sequence of matrices or CSR arrays to a Value object
}
