% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-models.R
\name{AttentionModel}
\alias{AttentionModel}
\title{Attention Model}
\usage{
AttentionModel(attention_dim, attention_span = NULL, attention_axis = NULL,
  init = init_glorot_uniform(), go_backwards = FALSE,
  enable_self_stabilization = TRUE, name = "")
}
\arguments{
\item{init}{(scalar or matrix or initializer, defaults to
init_glorot_uniform()) – initial value of weights W}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Attention Model
}
