% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-distributed.R
\name{block_momentum_distributed_learner}
\alias{block_momentum_distributed_learner}
\title{Block Momentum Distributed Learner}
\usage{
block_momentum_distributed_learner(learner, block_size,
  block_momentum_as_time_constant = NULL, use_nestrov_momentum = TRUE,
  reset_sgd_momentum_after_aggregation = TRUE, block_learning_rate = 1,
  distributed_after = 0)
}
\arguments{
\item{distributed_after}{}
}
\description{
Block Momentum Distributed Learner
}
