% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{CheckpointConfig}
\alias{CheckpointConfig}
\title{New Checkpoint Config}
\usage{
CheckpointConfig(filename, frequency = NULL, restore = TRUE,
  preserve_all = FALSE)
}
\arguments{
\item{:}{filename (str): checkpoint file name.}

\item{:}{frequency (int): checkpointing period (number samples between checkpoints). If `NULL`, no checkpointing takes place.
If ``sys.maxsize``, a single checkpoint is taken at the end of the training.}

\item{:}{restore (bool): flag, indicating whether to restore from available checkpoint before the start of the training}

\item{:}{preserve_all (bool): saves all checkpoints, using ``filename`` as prefix and checkpoint index as a suffix.}
}
\description{
New Checkpoint Config
}
