% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{classification_error}
\alias{classification_error}
\title{Classification Error Between Target and Predicted}
\usage{
classification_error(output_vector, target_vector, axis = -1, topN = 1,
  name = "")
}
\arguments{
\item{output_vector}{the output values of the network}

\item{target_vector}{one-hot encoding of target values}

\item{axis}{integer for axis along which the classification error is computed}

\item{topN}{integer}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
This operation computes the classification error.
}
\details{
It finds the index of the highest value in the output_vector and compares it
to the actual ground truth label (the index of the hot bit in the target
vector).

The result is a scalar (i.e., one by one matrix). This is often used as an
evaluation criterion.

It cannot be used as a training criterion though since the gradient is not
defined for it.
}
\references{
\url{https://www.cntk.ai/pythondocs/cntk.metrics.html#cntk.metrics.classification_error}
}
