% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{CloneMethod}
\alias{CloneMethod}
\title{CloneMethod}
\usage{
CloneMethod(value)
}
\arguments{
\item{value}{}
}
\description{
Object descibing different ways how clone() works.
}
\details{
****** Attributes: ******

clone = 'clone' - New learnable parameters are created and initialized with
the current values of the corresponding parameters of the Function being
cloned

freeze = 'freeze' - Parameters are cloned and made immutable; i.e. Constants
in the new clone (e.g. for use as a fixed feature extractor)

share = 'share' - Parameters are shared between the Function being cloned
and the new clone
}
