% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-distributed.R
\name{comm_barrier}
\alias{comm_barrier}
\title{Communicator Barrier}
\usage{
comm_barrier(communicator)
}
\arguments{
\item{communicator}{- communicator instance on which to call operation}
}
\description{
Communicator Barrier
}
