% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-distributed.R
\name{comm_current_worker}
\alias{comm_current_worker}
\title{Communicator Current Worker}
\usage{
comm_current_worker(communicator)
}
\arguments{
\item{communicator}{- communicator instance on which to call operation}
}
\description{
Communicator Current Worker
}
