% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-distributed.R
\name{comm_is_main}
\alias{comm_is_main}
\title{Communicator Is On Main Node}
\usage{
comm_is_main(communicator)
}
\arguments{
\item{communicator}{- communicator instance on which to call operation}
}
\description{
Communicator Is On Main Node
}
