% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-distributed.R
\name{comm_workers}
\alias{comm_workers}
\title{Communicator Workers}
\usage{
comm_workers(communicator)
}
\arguments{
\item{communicator}{- communicator instance on which to call operation}
}
\description{
Communicator Workers
}
