% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-distributed.R
\name{Communicator}
\alias{Communicator}
\title{New Communicator}
\usage{
Communicator(...)
}
\arguments{
\item{...}{- constructor args}
}
\description{
A communicator interface exposing communication primitives that serve as
building blocks for distributed training.
}
