% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Constant}
\alias{Constant}
\title{Constant}
\usage{
Constant(value = NULL, shape = NULL, dtype = "float32", device = NULL,
  name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{dtype}{- data type to be used ("float32", "float64", or "auto")}

\item{device}{- instance of DeviceDescriptor}

\item{name}{}
}
\description{
A constant value. It can be a scalar, vector, matrix, or tensor of floating
point numbers that cannot be modified.  A Constant is a Variable and
therefore inherits all its methods.
}
\details{
****** Properties: ******

value
}
