% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-layers.R
\name{Convolution3D}
\alias{Convolution3D}
\title{Convolution3D}
\usage{
Convolution3D(filter_shape, num_filters = NULL,
  activation = activation_identity, init = init_glorot_uniform(),
  pad = FALSE, strides = 1, bias = TRUE, init_bias = 0,
  reduction_rank = 1, name = "")
}
\arguments{
\item{name}{string for the name of the function instance in the network}
}
\description{
Layer factory function to create a 3D convolution layer with optional
non-linearity. Same as Convolution() except that filter_shape is verified to
be 3-dimensional. See Convolution() for extensive documentation.
}
