% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-layers.R
\name{ConvolutionTranspose2D}
\alias{ConvolutionTranspose2D}
\title{ConvolutionTranspose2D}
\usage{
ConvolutionTranspose2D(filter_shape, num_filters = NULL,
  activation = activation_identity, init = init_glorot_uniform(),
  pad = FALSE, strides = 1, bias = TRUE, init_bias = 0,
  output_shape = NULL, name = "")
}
\arguments{
\item{name}{string for the name of the function instance in the network}
}
\description{
Layer factory function to create a 2D convolution transpose layer with
optional non-linearity. Same as ConvolutionTranspose() except that
filter_shape is verified to be 2-dimensional. See ConvolutionTranspose() for
extensive documentation.
}
