% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{CrossValidationConfig}
\alias{CrossValidationConfig}
\title{New Cross Validation Configuration}
\usage{
CrossValidationConfig(minibatch_source = NULL, frequency = NULL,
  minibatch_size = 32, callback = NULL, max_samples = NULL,
  model_inputs_to_streams = NULL, criterion = NULL, source = NULL,
  mb_size = NULL)
}
\arguments{
\item{mb_size}{}
}
\description{
New Cross Validation Configuration
}
