% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_assign}
\alias{ct_assign}
\title{Assign Crosstalk Value}
\usage{
ct_assign(ct, name, value = NULL, load = FALSE, load_name = NULL)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}

\item{name}{string (optional) the name of the Function instance in the network}

\item{load_name}{}
}
\description{
Assign Crosstalk Value
}
