% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_compare}
\alias{ct_compare}
\title{Compare Crosstalk Var}
\usage{
ct_compare(ct, name, compare_name = NULL, rtol = 1e-05, atol = 1e-08,
  equal_nan = FALSE)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}

\item{name}{string (optional) the name of the Function instance in the network}

\item{equal_nan}{}
}
\description{
Compare Crosstalk Var
}
