% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_fetch}
\alias{ct_fetch}
\title{Fetch Crosstalk Var}
\usage{
ct_fetch(ct, name, save = FALSE)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}

\item{name}{string (optional) the name of the Function instance in the network}

\item{save}{}
}
\description{
Fetch Crosstalk Var
}
