% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_is_param}
\alias{ct_is_param}
\title{CNTKCrosstalk Var is Parameter}
\usage{
ct_is_param(cct, name)
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
CNTKCrosstalk Var is Parameter
}
