% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_load}
\alias{ct_load}
\title{Load Crosstalk Vars}
\usage{
ct_load(ct, names)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Load Crosstalk Vars
}
