% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_load_raw_value}
\alias{ct_load_raw_value}
\title{Load Crosstalk Raw Value}
\usage{
ct_load_raw_value(ct, name)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Load Crosstalk Raw Value
}
