% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_next_pass}
\alias{ct_next_pass}
\title{Next Crosstalk Pass}
\usage{
ct_next_pass(ct)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}
}
\description{
Next Crosstalk Pass
}
