% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_register_funcs}
\alias{ct_register_funcs}
\title{Register Crosstalk Var Type Getter/Setters}
\usage{
ct_register_funcs(ct, var_type, setter = NULL, getter = NULL)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}

\item{getter}{}
}
\description{
Register Crosstalk Var Type Getter/Setters
}
