% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_save}
\alias{ct_save}
\title{Save Crosstalk Vars}
\usage{
ct_save(ct, names)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Save Crosstalk Vars
}
