% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_save_all}
\alias{ct_save_all}
\title{Save All Crosstalk Vars}
\usage{
ct_save_all(ct)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}
}
\description{
Save All Crosstalk Vars
}
