% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_set_workdir}
\alias{ct_set_workdir}
\title{Set Crosstalk Working Directory}
\usage{
ct_set_workdir(ct, dir)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}

\item{dir}{}
}
\description{
Set Crosstalk Working Directory
}
