% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{ct_watch}
\alias{ct_watch}
\title{Watch Crosstalk Variable}
\usage{
ct_watch(ct, var, name, var_type = NULL, attr = NULL)
}
\arguments{
\item{ct}{Crosstalk instance on which to perform the operation}

\item{name}{string (optional) the name of the Function instance in the network}

\item{attr}{}
}
\description{
Watch Crosstalk Variable
}
