% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{CTFDeserializer}
\alias{CTFDeserializer}
\title{CTFDeserializer}
\usage{
CTFDeserializer(filename, streams)
}
\description{
Configures the CNTK text-format reader that reads text-based files.
}
\details{
Form: `[Sequence_Id] (Sample)+` where `Sample=|Input_Name (Value )*`
}
