% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-distributed.R
\name{data_parallel_distributed_learner}
\alias{data_parallel_distributed_learner}
\title{New Data Parallel Distributed Learner}
\usage{
data_parallel_distributed_learner(learner, distributed_after = 0,
  num_quantization_bits = 32, async_parameter_update = FALSE)
}
\arguments{
\item{async_parameter_update}{}
}
\description{
New Data Parallel Distributed Learner
}
