% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging-debug.R
\name{debug_model}
\alias{debug_model}
\title{Debug Model}
\usage{
debug_model(model, in_stream = sys$stdin, out = sys$stdout,
  exit_func = sys$exit)
}
\arguments{
\item{exit_func}{}
}
\description{
Debug Model
}
