% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-sequence.R
\name{Delay}
\alias{Delay}
\title{Delay}
\usage{
Delay(T = 1, initial_state = 0, name = "")
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}

\item{init}{(scalar or matrix or initializer, defaults to
init_glorot_uniform()) – initial value of weights Wial_state}
}
\description{
Delay
}
