% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-distributed.R
\name{DistributedLearner}
\alias{DistributedLearner}
\title{Distributed Learner}
\usage{
DistributedLearner(...)
}
\arguments{
\item{...}{}
}
\description{
A distributed learner that handles data like gradients/momentums across multiple MPI workers
}
\details{
****** Properties: ******

total_number_of_samples_seen
}
