% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_test_minibatch}
\alias{eval_test_minibatch}
\title{Test Evaluator Minibatch}
\usage{
eval_test_minibatch(evaluator, arguments, device = NULL)
}
\arguments{
\item{evaluator}{the Evaluator instance on which to perform the operation}

\item{arguments}{named list of input variable names to input data or if node
has a unique input, arguments is mapped to this input. For nodes with more
than one input, only a named list is allowed.}

\item{device}{- instance of DeviceDescriptor}
}
\description{
Test Evaluator Minibatch
}
