% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{Evaluator}
\alias{Evaluator}
\title{Evaluator}
\usage{
Evaluator(eval_function, progress_writers = NULL)
}
\arguments{
\item{progress_writers}{}
}
\description{
Class for evaluation of minibatches against the specified evaluation
function.
}
\details{
****** Attributes: ******

evaluation_function

****** Associated Functions: ******

summarize_test_progress(evaluator)

eval_test_minibatch(evaluator, arguments, device = NULL)
}
