% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk-cntk.R
\name{find_func_param}
\alias{find_func_param}
\title{Find Parameter in Function - CNTKCrosstalk}
\usage{
find_func_param(func, name = NULL, shape = NULL, allow_not_found = FALSE)
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}

\item{shape}{- list of ints representing tensor shape}

\item{allow_not_found}{}
}
\description{
Find Parameter in Function - CNTKCrosstalk
}
