% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-sequence.R
\name{Fold}
\alias{Fold}
\title{Fold}
\usage{
Fold(folder_function, go_backwards, initial_state = 0,
  return_full_state = FALSE, name = "")
}
\arguments{
\item{folder_function}{}

\item{name}{string (optional) the name of the Function instance in the network}

\item{init}{(scalar or matrix or initializer, defaults to
init_glorot_uniform()) – initial value of weights Wial_state}
}
\description{
Fold
}
