% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-higher_order_layers.R
\name{For}
\alias{For}
\title{CNTK Function For Loop Construct}
\usage{
For(range, constructor, name = "")
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
CNTK Function For Loop Construct
}
