% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-blocks.R
\name{ForwardDeclaration}
\alias{ForwardDeclaration}
\title{ForwardDeclaration}
\usage{
ForwardDeclaration(name = "forward_declaration")
}
\description{
Helper for recurrent network declarations. Returns a placeholder variable
with an added method resolve_to() to be called at the end to close the loop.
This is used for explicit graph building with recurrent connections.
}
