% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_backward}
\alias{func_backward}
\title{Propogate Function Backward}
\usage{
func_backward(func, state, root_gradients, variables, as_matrix = TRUE)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{as_matrix}{- whether to return as an R matrix. Defualt TRUE. Otherwise
returns as Python CNTK value which avoids costly conversions}
}
\description{
Backpropagates supplied root_gradients for one or more of the output
variables of the Function, to calculate gradients with respect to variables.
Formally, multiplies the values of root_gradients by the Jacobian of the
Function and returns the subset of the output that corresponds to variables.
}
