% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_clone}
\alias{func_clone}
\title{Clone Function}
\usage{
func_clone(func, method, substitutions = NULL)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{substitutions}{}
}
\description{
Clones the function. The parameters of the Function are either cloned,
shared or frozen as specified by the method argument and any variable
substitutions requested are applied in the cloned Function instance.
}
