% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_eval}
\alias{func_eval}
\title{Evaluate Function}
\usage{
func_eval(func, arguments = NULL, outputs = NULL, device = NULL,
  as_matrix = TRUE)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{device}{- instance of DeviceDescriptor}

\item{as_matrix}{- whether to return as an R matrix. Defualt TRUE. Otherwise
returns as Python CNTK value which avoids costly conversions}
}
\description{
Evaluate the Function’s outputs using the specified arguments as input.
}
