% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_find_all_with_name}
\alias{func_find_all_with_name}
\title{Find All Functions With Name}
\usage{
func_find_all_with_name(func, name, depth = 0)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{depth}{}
}
\description{
Returns a list of primitive function with name in the graph starting from
this node. Throws an exception if name occurs multiple times. If you expect
only one function to be returned, use find_by_name().
}
