% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_find_by_name}
\alias{func_find_by_name}
\title{Find Function By Name}
\usage{
func_find_by_name(func, name, depth = 0)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{depth}{}
}
\description{
Returns a primitive function with name in the graph starting from this node.
Throws an exception if name occurs multiple times. If you expect multiple
functions to be returned, use find_all_with_name().
}
