% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_forward}
\alias{func_forward}
\title{Compute Function Forward}
\usage{
func_forward(func, arguments, outputs = NULL, keep_for_backward = NULL,
  device = NULL, as_matrix = TRUE)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{device}{- instance of DeviceDescriptor}

\item{as_matrix}{- whether to return as an R matrix. Defualt TRUE. Otherwise
returns as Python CNTK value which avoids costly conversions}
}
\description{
Computes the values of speficied variables in outputs, using values provided
in arguments that correspond to each input Variable of the function (i.e.
those that have is_input = True).
}
