% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_grad}
\alias{func_grad}
\title{Compute Function Gradient}
\usage{
func_grad(func, at, wrt = NULL, outputs = NULL, device = NULL,
  as_matrix = TRUE, grad_root = NULL)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{at}{- mapping of the Function’s arguments to values}

\item{wrt}{- list of Variables with respect to which the gradient will be
computed. If omitted, the gradients with respect to all arguments of this
Function that need gradient will be computed.}

\item{device}{- instance of DeviceDescriptor}

\item{as_matrix}{- whether to return as an R matrix. Defualt TRUE. Otherwise
returns as Python CNTK value which avoids costly conversions}

\item{grad_root}{- specify the root of gradients calculation. If not
specified, the output of this function will be used as gradient root.}
}
\description{
Computes the gradient of this Function at location at with respect to wrt.
The Function must have a single output.
}
