% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_load}
\alias{func_load}
\title{Load Function Model}
\usage{
func_load(model, device = NULL)
}
\arguments{
\item{device}{- instance of DeviceDescriptor}
}
\description{
Load the model, that has been saved using save().
}
