% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_replace_placeholder}
\alias{func_replace_placeholder}
\title{Replace Function Placeholder}
\usage{
func_replace_placeholder(func, substitution)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{substitution}{- `Variable` that will replace the placeholder}
}
\description{
In-place replace the only placeholder in the function graph with the
specified substitution.
}
