% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_replace_placeholders}
\alias{func_replace_placeholders}
\title{Replace Function Placeholders}
\usage{
func_replace_placeholders(func, substitutions)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{substitutions}{- dict mapping placeholders to variables}
}
\description{
In-place replace specified placeholders in the Function graph with the
specified replacements in the map.
}
