% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_restore}
\alias{func_restore}
\title{Restore Function}
\usage{
func_restore(func, filename)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{filename}{}
}
\description{
Restore model parameters from saved model file.
}
