% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_save}
\alias{func_save}
\title{Save Function}
\usage{
func_save(func, filename)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{filename}{}
}
\description{
Save this function graph into a model file using protobuf-based serialization.
}
\details{
Use comm_is_main() to gate your call to save() in distributed environment.
}
