% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_set_attribute}
\alias{func_set_attribute}
\title{Set Function Attribute}
\usage{
func_set_attribute(func, name, value)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{value}{}
}
\description{
Set Function Attribute
}
