% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{func_test}
\alias{func_test}
\title{Test Function Model}
\usage{
func_test(func, minibatch_source, minibatch_size = 32, streams = NULL,
  model_inputs_to_streams = NULL, callbacks = NULL)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{minibatch_source}{- minibatch source for the test data}

\item{minibatch_size}{(minibatch_size_schedule or int) – minibatch size for
evaluation}

\item{streams}{(list) - the streams of the minibatch_source in argument
order}

\item{model_inputs_to_streams}{(dict or named list) - mapping between input
variables and #' input streams}

\item{callbacks}{(progress writer or list of them) – optionally, list of
progress writers from cntk.logging to automatically track training progress.}
}
\description{
Measures the performance of a model, given by its criterion function, in the
form of average metric value (or loss if model has only one output) on a set
of data.
}
