% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-distributed.R
\name{get_communicator}
\alias{get_communicator}
\title{Get Distributed Communicator}
\usage{
get_communicator(distributed_learner)
}
\arguments{
\item{distributed_learner}{}
}
\description{
Get Distributed Communicator
}
