% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_options.R
\name{get_default_override}
\alias{get_default_override}
\title{Get Default Override}
\usage{
get_default_override(function_or_class, ...)
}
\arguments{
\item{...}{}
}
\description{
Get Default Override
}
