% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_minibatch_checkpoint_state}
\alias{get_minibatch_checkpoint_state}
\title{get_minibatch_checkpoint_state}
\usage{
get_minibatch_checkpoint_state(mb_source)
}
\description{
Returns a dictionary describing the current state of the minibatch source.
Needs to be overwritten if the state of the minibatch source needs to be
stored to and later restored from the checkpoint.
}
