% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{get_static_axis_index}
\alias{get_static_axis_index}
\title{get_static_axis_index}
\usage{
get_static_axis_index(ax, checked = TRUE)
}
\arguments{
\item{checked}{}
}
\description{
Returns the integer with which the static axis is defined. For example, 0 =
first axis, 1 = second axis, etc.
}
