% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-layers.R
\name{GlobalAveragePooling}
\alias{GlobalAveragePooling}
\title{GlobalAveragePooling}
\usage{
GlobalAveragePooling(name = "")
}
\description{
Layer factory function to create a global average-pooling layer.
}
\details{
The global average-pooling operation computes the element-wise mean over all
items on an N-dimensional grid, such as an image.

This operation is the same as applying reduce_mean() to all grid dimensions.
}
