% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-graph.R
\name{graph_depth_first_search}
\alias{graph_depth_first_search}
\title{Graph - Depth First Search}
\usage{
graph_depth_first_search(root, visitor, depth = 0)
}
\arguments{
\item{depth}{}
}
\description{
Graph - Depth First Search
}
