% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-graph.R
\name{graph_find_by_name}
\alias{graph_find_by_name}
\title{Graph - Find By Name}
\usage{
graph_find_by_name(node, node_name, depth = 0)
}
\arguments{
\item{node}{}

\item{depth}{}
}
\description{
Graph - Find By Name
}
