% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{HTKFeatureDeserializer}
\alias{HTKFeatureDeserializer}
\title{HTKFeatureDeserializer}
\usage{
HTKFeatureDeserializer(streams)
}
\arguments{
\item{streams}{any dictionary-like object that contains a mapping from
stream names to StreamDef objects. Each StreamDef object configures a label
stream.}
}
\description{
Configures the HTK feature reader that reads speech data from scp files.
}
