% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{HTKMLFDeserializer}
\alias{HTKMLFDeserializer}
\title{Base64ImageDeserializer}
\usage{
HTKMLFDeserializer(label_mapping_file, streams, phoneBoundaries = FALSE)
}
\arguments{
\item{label_mapping_file}{label mpaping file}

\item{streams}{any dictionary-like object that contains a mapping from
stream names to StreamDef objects. Each StreamDef object configures a label
stream.}

\item{phoneBoundaries}{phone boundaries}
}
\description{
Configures an HTK label reader that reads speech HTK format MLF (Master
Label File)
}
