% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{init_he_uniform}
\alias{init_he_uniform}
\title{He Uniform Initializer}
\usage{
init_he_uniform(scale = 1, output_rank = 2147483647,
  filter_rank = 2147483647, seed = NULL)
}
\arguments{
\item{seed}{}
}
\description{
He Uniform Initializer
}
