% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{init_normal}
\alias{init_normal}
\title{Normal Initializer}
\usage{
init_normal(scale = 1, output_rank = 2147483647, filter_rank = 2147483647,
  seed = NULL)
}
\arguments{
\item{seed}{}
}
\description{
Normal Initializer
}
