% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{init_truncated_normal}
\alias{init_truncated_normal}
\title{Truncated Normal Initializer}
\usage{
init_truncated_normal(stdev, seed = NULL)
}
\arguments{
\item{stdev}{(float) standard deviation}

\item{seed}{(int) random seed}
}
\description{
Truncated Normal Initializer
}
