% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{init_uniform}
\alias{init_uniform}
\title{Uniform Initializer}
\usage{
init_uniform(scale, seed = NULL)
}
\arguments{
\item{scale}{(float) scale}

\item{seed}{(int) random seed}
}
\description{
Uniform Initializer
}
