% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{init_with_rank}
\alias{init_with_rank}
\title{Initializer With Rank}
\usage{
init_with_rank(initializer, output_rank = NULL, filter_rank = NULL)
}
\arguments{
\item{output_rank}{(int) output rank}

\item{filter_rank}{(int) filter rank}

\item{init}{- Initializer to be overriden}
}
\description{
override output_rank and filter_rank specification in a random initializer
constructed without an explicit output_rank and filter_rank specification
}
