% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_installer.R
\name{install_samples}
\alias{install_samples}
\title{Install Samples}
\usage{
install_samples(url = NULL, directory = NULL, quiet = FALSE)
}
\arguments{
\item{quiet}{}
}
\description{
Fetch the CNTK samples from a URL, extract to local directory, and install
Python package requirements.
}
