% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{data}
\name{IO_INFINITELY_REPEAT}
\alias{IO_INFINITELY_REPEAT}
\title{IO_INFINITELY_REPEAT}
\format{An object of class \code{integer} of length 1.}
\usage{
IO_INFINITELY_REPEAT
}
\description{
Constant used to specify a minibatch scheduling unit to equal the size of
the full data sweep.
}
\keyword{datasets}
