% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{Learner}
\alias{Learner}
\title{Learner}
\usage{
Learner(parameters, learning_rate_schedule)
}
\arguments{
\item{parameters}{– list of network parameters list of parameter associated
with this learner}

\item{learning_rate_schedule}{get_learning_rate(learner)}
}
\description{
Abstraction for learning a subset of parameters of a learnable function
using first order gradient values. For example momentum, AdaGrad, RMSProp,
etc. are different types of learners with their own algorithms for learning
parameter values using first order gradients. To instantiate a concrete
learner, use the factory methods in this module.
}
