% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{log_number_of_parameters}
\alias{log_number_of_parameters}
\title{Log Number of Parameters}
\usage{
log_number_of_parameters(model, trace_level = 0)
}
\arguments{
\item{trace_level}{}
}
\description{
Log Number of Parameters
}
