% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_binary_cross_entropy}
\alias{loss_binary_cross_entropy}
\title{Binary Cross Entropy Loss}
\usage{
loss_binary_cross_entropy(output, target, name = "")
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Binary Cross Entropy Loss
}
