% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_cosine_distance}
\alias{loss_cosine_distance}
\title{Cosine Distance Loss}
\usage{
loss_cosine_distance(x, y, name = "")
}
\arguments{
\item{x}{}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Cosine Distance Loss
}
