% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_cosine_distance_negative_samples}
\alias{loss_cosine_distance_negative_samples}
\title{Cosine Distance NEgative Samples Loss}
\usage{
loss_cosine_distance_negative_samples(x, y, shift, num_negative_samples,
  name = "")
}
\arguments{
\item{x}{}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Cosine Distance NEgative Samples Loss
}
