% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_lambda_rank}
\alias{loss_lambda_rank}
\title{Lambda Rank Loss}
\usage{
loss_lambda_rank(output, gain, group, name = "")
}
\arguments{
\item{output}{}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Lambda Rank Loss
}
