% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_squared_error}
\alias{loss_squared_error}
\title{Squared Error Loss}
\usage{
loss_squared_error(output, target, name = "")
}
\arguments{
\item{output}{}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Squared Error Loss
}
