% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{loss_weighted_binary_cross_entropy}
\alias{loss_weighted_binary_cross_entropy}
\title{Weighted Binary Cross Entropy Loss}
\usage{
loss_weighted_binary_cross_entropy(output, target, weight, name = "")
}
\arguments{
\item{output}{}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Weighted Binary Cross Entropy Loss
}
