% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mb_as_sequences}
\alias{mb_as_sequences}
\title{Minibatch Data As Sequences}
\usage{
mb_as_sequences(minibatch_data, variable = NULL)
}
\arguments{
\item{minibatch_data}{the MinibatchData instance on which to perform the
operation}
}
\value{
a list of matrices if dense, otherwise a SciPy CSR array
}
\description{
Convert the value of this minibatch instance to a sequence of NumPy arrays
that have their masked entries removed.
}
