% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mb_stream_info}
\alias{mb_stream_info}
\title{Minibatch Stream Info}
\usage{
mb_stream_info(mb_source, name)
}
\arguments{
\item{mb_source}{minibatch source on which to perform the operation}

\item{name}{string (optional) the name of the Function instance in the
network}
}
\description{
Gets the description of the stream with given name. Throws an exception if
there are none or multiple streams with this same name.
}
