% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mb_stream_infos}
\alias{mb_stream_infos}
\title{Minibatch Stream Infos}
\usage{
mb_stream_infos(mb_source)
}
\arguments{
\item{mb_source}{minibatch source on which to perform the operation}
}
\description{
Function to be implemented by the user.
}
