% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{MinibatchData}
\alias{MinibatchData}
\title{MinibatchData}
\usage{
MinibatchData(value, num_sequences, num_samples, sweep_end)
}
\description{
Holds a minibatch of input data. This is never directly created, but only
returned by `MinibatchSource` instances.
}
\details{
****** Attributes: ******

data

end_of_sweep

is_sparse

mask

num_samples

num_sequences

shape

****** Associated Functions: ******

mb_as_sequences(variable = NULL)
}
