% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{MinibatchSource}
\alias{MinibatchSource}
\title{MinibatchSource}
\usage{
MinibatchSource(deserializers, max_samples = IO_INFINITELY_REPEAT,
  max_sweeps = IO_INFINITELY_REPEAT,
  randomization_window_in_chunks = IO_DEFAULT_RANDOMIZATION_WINDOW_IN_CHUNKS,
  randomization_window_in_samples = 0, randomization_seed = 0,
  trace_level = get_logging_trace_level(),
  multithreaded_deserializer = NULL, frame_mode = FALSE,
  truncation_length = 0, randomize = TRUE)
}
\arguments{
\item{randomize}{}
}
\description{
Source of minibatch data
}
\details{
****** Attributes: ******

current_position

is_distributed

streams

****** Associated Functions: ******

get_minibatch_checkpoint_state(minibatch_source)

next_minibatch(minibatch_source, minibatch_size_in_samples,
input_map = NULL, device = NULL, num_data_partitions = NULL,
partition_index = NULL)

restore_mb_from_checkpoint(minibatch_source, checkpoint)

mb_stream_info(minibatch_source, name)
mb_stream_infos(minibatch_source)
}
