% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{native_user_function}
\alias{native_user_function}
\title{Create Native UserFunction}
\usage{
native_user_function(op_id, operands, attributes = NULL,
  user_function_instance_name = "")
}
\arguments{
\item{user_function_instance_name}{}
}
\description{
Creates an instance of a user-defined Function previously registered using
the ‘register_native_user_function’ method.
}
