% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{NDArrayView}
\alias{NDArrayView}
\title{NDArrayView}
\usage{
NDArrayView(shape, dtype, device = NULL)
}
\arguments{
\item{shape}{list - shape of the data}

\item{dtype}{"float32" or "float64" - data type}

\item{device}{DeviceDescriptor - device this value should be put on}
}
\description{
Creates an empty dense internal data representation of a Value object. To create an NDArrayView from a NumPy array, use from_dense(). To create an NDArrayView from a sparse array, use from_csr().
}
