% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ndcg_at_1}
\alias{ndcg_at_1}
\title{NDCG at 1}
\usage{
ndcg_at_1(output, gain, group, name = "")
}
\arguments{
\item{name}{}
}
\description{
Groups samples according to group, sorts them within each group based on
output and computes the Normalized Discounted Cumulative Gain (NDCG) at 1
for each group. Concretely, the NDCG at 1 is:
}
\details{
\(\mathrm{NDCG_1} = \frac{gain_{(1)}}{\max_i gain_i}\)

where gain(1)gain(1) means the gain of the first ranked sample.

Samples in the same group must appear in order of decreasing gain.

It returns the average NDCG at 1 across all the groups in the minibatch
multiplied by 100 times the number of samples in the minibatch.

This is a forward-only operation, there is no gradient for it.
}
