% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{on_train_cross_validation_end}
\alias{on_train_cross_validation_end}
\title{On Training Cross Validation End}
\usage{
on_train_cross_validation_end(sess, index, average_error, num_samples,
  num_minibatches)
}
\arguments{
\item{sess}{- session instance on which to perform the operation}

\item{num_minibatches}{}
}
\description{
On Training Cross Validation End
}
