% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_argmin}
\alias{op_argmin}
\title{Argmin Across Axis}
\usage{
op_argmin(x, axis = NULL, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{axis}{- axis across which to perform operation}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Argmin Across Axis
}
