% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_associative_multi_arg}
\alias{op_associative_multi_arg}
\title{Associative Multi-Arg}
\usage{
op_associative_multi_arg(f)
}
\arguments{
\item{f}{}
}
\description{
The output of this operation is the result of an operation (plus,
log_add_exp, element_times, element_max, element_min) of two or more input
tensors. Broadcasting is supported.
}
