% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_batch_normalization}
\alias{op_batch_normalization}
\title{Batch Normalization}
\usage{
op_batch_normalization(operand, scale, bias, running_mean, running_inv_std,
  spatial, normalization_time_constant = 5000, blent_time_constant = 0,
  epsilon = 1e-05, use_cudnn_engine = FALSE, name = "",
  running_count = NULL)
}
\arguments{
\item{bias}{(bool) – whether to include bias}

\item{epsilon}{(float, default 0.00001) - added to avoid division by 0}

\item{name}{(str) - the name of the Function instance in the network}

\item{running_count}{}
}
\description{
Normalizes layer outputs for every minibatch for each output (feature)
independently and applies affine transformation to preserve representation
of the layer.
}
