% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_ceil}
\alias{op_ceil}
\title{Ceiling}
\usage{
op_ceil(arg, name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
The output of this operation is the element wise value rounded to the
smallest integer greater than or equal to the input.
}
