% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_clip}
\alias{op_clip}
\title{Clip}
\usage{
op_clip(x, min_value, max_value, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Computes a tensor with all of its values clipped to fall between min_value
and max_value, i.e. min(max(x, min_value), max_value).  The output tensor
has the same shape as x.
}
