% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_combine}
\alias{op_combine}
\title{Combine}
\usage{
op_combine(operands, name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Create a new Function instance which just combines the outputs of the
specified list of ‘operands’ Functions such that the ‘Outputs’ of the new
‘Function’ are union of the ‘Outputs’ of each of the specified ‘operands’
Functions. E.g., when creating a classification model, typically the
CrossEntropy loss Function and the ClassificationError Function comprise the
two roots of the computation graph which can be combined to create a single
Function with 2 outputs; viz. CrossEntropy loss and ClassificationError
output.
}
