% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_constant}
\alias{op_constant}
\title{It creates a constant tensor initialized from a numpy array}
\usage{
op_constant(value = NULL, shape = NULL, dtype = "auto", name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{dtype}{- data type to be used ("float32", "float64", or "auto")}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
It creates a constant tensor initialized from a numpy array
}
