% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_cos}
\alias{op_cos}
\title{Element-wise Cosine}
\usage{
op_cos(x, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Computes the element-wise cosine of x: The output tensor has the same shape
as x.
}
