% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_dim_mean_variance_normalize}
\alias{op_dim_mean_variance_normalize}
\title{Per-dimension Mean-variance Normalization}
\usage{
op_dim_mean_variance_normalize(operand, mean, inv_stddev, name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Per-dimension Mean-variance Normalization
}
