% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_element_select}
\alias{op_element_select}
\title{Element Select}
\usage{
op_element_select(flag, value_if_true, value_if_false, name = "")
}
\arguments{
\item{flag}{}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Element Select
}
