% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_element_times}
\alias{op_element_times}
\title{Element Times}
\usage{
op_element_times(left, right, name = "")
}
\arguments{
\item{left}{- left side tensor}

\item{right}{- right side tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Element Times
}
