% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_equal}
\alias{op_equal}
\title{Equal Comparison}
\usage{
op_equal(left, right, name = "")
}
\arguments{
\item{left}{- left side tensor}

\item{right}{- right side tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Equal Comparison
}
