% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_forward_backward}
\alias{op_forward_backward}
\title{Forward-Backward}
\usage{
op_forward_backward(graph, features, blank_token_id, delay_constraint = -1,
  name = "")
}
\arguments{
\item{graph}{}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Forward-Backward
}
