% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_greater_equal}
\alias{op_greater_equal}
\title{Element-wise Greater Equal Comparison}
\usage{
op_greater_equal(left, right, name = "")
}
\arguments{
\item{left}{- left side tensor}

\item{right}{- right side tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Element-wise Greater Equal Comparison
}
