% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_input_variable}
\alias{op_input_variable}
\title{Create input for network}
\usage{
op_input_variable(shape, needs_gradient = FALSE, is_sparse = FALSE,
  dynamic_axes = c(get_default_batch_axis()), name = "")
}
\arguments{
\item{shape}{integer vector for dimensions of input tensor}

\item{needs_gradient}{logical whether to conduct backprop on the tensor}

\item{is_sparse}{logical whether variable is sparse}

\item{dynamic_axes}{list of dynamic axis (only a single axis can be dynamic, i.e., either batch axis or time axis)}
}
\value{
Variable \url{https://www.cntk.ai/pythondocs/cntk.variables.html#cntk.variables.Variable}
}
\description{
It creates an input in the network: a place where data, such as features and labels, should be provided.
}
\references{
\url{https://www.cntk.ai/pythondocs/cntk.ops.html#cntk.ops.input_variable}
}
