% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_labels_to_graph}
\alias{op_labels_to_graph}
\title{Labels To Graph}
\usage{
op_labels_to_graph(labels, name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Labels To Graph
}
