% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_leaky_relu}
\alias{op_leaky_relu}
\title{Leaky Relu}
\usage{
op_leaky_relu(x, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Leaky Relu
}
