% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_less}
\alias{op_less}
\title{Element-wise Less Comparison}
\usage{
op_less(left, right, name = "")
}
\arguments{
\item{left}{- left side tensor}

\item{right}{- right side tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Element-wise Less Comparison
}
