% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_log}
\alias{op_log}
\title{Element-wise Natural Log}
\usage{
op_log(x, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Element-wise Natural Log
}
