% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_log_add_exp}
\alias{op_log_add_exp}
\title{Log Add Exp}
\usage{
op_log_add_exp(left, right, name = "")
}
\arguments{
\item{left}{- left side tensor}

\item{right}{- right side tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Calculates the log of the sum of the exponentials of the two or more input
tensors. It supports broadcasting.
}
