% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_minus}
\alias{op_minus}
\title{Minus}
\usage{
op_minus(left, right, name = "")
}
\arguments{
\item{left}{- left side tensor}

\item{right}{- right side tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Minus
}
