% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_one_hot}
\alias{op_one_hot}
\title{Create One-Hot Encoding}
\usage{
op_one_hot(x, num_classes, sparse_output = FALSE, axis = -1, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{axis}{- axis across which to perform operation}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Create One-Hot Encoding
}
