% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_optimized_rnnstack}
\alias{op_optimized_rnnstack}
\title{Optimized RNN Stack}
\usage{
op_optimized_rnnstack(operand, weights, hidden_size, num_layers,
  bidirectional = FALSE, recurrent_op = "lstm", name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Optimized RNN Stack
}
