% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_output_variable}
\alias{op_output_variable}
\title{Output Variable}
\usage{
op_output_variable(shape, dtype, dynamic_axes, needs_gradient = TRUE,
  name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{dtype}{- data type to be used ("float32", "float64", or "auto")}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Output Variable
}
