% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_param_relu}
\alias{op_param_relu}
\title{Parametric ReLU}
\usage{
op_param_relu(alpha, x, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Parametric ReLU
}
