% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_parameter}
\alias{op_parameter}
\title{Parameter}
\usage{
op_parameter(shape = NULL, init = NULL, dtype = "auto", device = NULL,
  name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{init}{(scalar or matrix or initializer, defaults to
init_glorot_uniform()) – initial value of weights W}

\item{dtype}{- data type to be used ("float32", "float64", or "auto")}

\item{device}{- instance of DeviceDescriptor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Creates a parameter tensor
}
