% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_placeholder}
\alias{op_placeholder}
\title{Placeholder}
\usage{
op_placeholder(shape = NULL, dynamic_axes = NULL, name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Placeholder
}
