% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_plus}
\alias{op_plus}
\title{Addition of Two Tensors}
\usage{
op_plus(left, right, name = "")
}
\arguments{
\item{left}{- left side tensor}

\item{right}{- right side tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Addition of Two Tensors
}
