% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_pooling}
\alias{op_pooling}
\title{Pooling}
\usage{
op_pooling(operand, pooling_type, pooling_window_shape, strides = c(1),
  auto_padding = c(FALSE), ceil_out_dim = FALSE, include_pad = FALSE,
  name = "")
}
\arguments{
\item{strides}{(int or tuple of ints, defaults to 1) – stride of the
operation. Use a list of ints to specify a per-axis value.}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Pooling
}
