% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_pow}
\alias{op_pow}
\title{Power Computation}
\usage{
op_pow(base, exponent, name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Power Computation
}
