% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_random_sample}
\alias{op_random_sample}
\title{Random Sample}
\usage{
op_random_sample(weights, num_samples, allow_duplicates, seed = 4294967293,
  name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Random Sample
}
