% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_random_sample_inclusion_frequency}
\alias{op_random_sample_inclusion_frequency}
\title{Random Sample Inclusion Frequency}
\usage{
op_random_sample_inclusion_frequency(weights, num_samples, allow_duplicates,
  seed = 4294967293, name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Random Sample Inclusion Frequency
}
