% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_reciprocal}
\alias{op_reciprocal}
\title{Element-wise Reciprocal}
\usage{
op_reciprocal(x, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Element-wise Reciprocal
}
