% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_reconcile_dynamic_axes}
\alias{op_reconcile_dynamic_axes}
\title{Reconcile Dynamic Axes}
\usage{
op_reconcile_dynamic_axes(x, dynamic_axes_as, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Reconcile Dynamic Axes
}
