% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_reduce_prod}
\alias{op_reduce_prod}
\title{Reduce Prod Across Axis}
\usage{
op_reduce_prod(x, axis = NULL, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{axis}{- axis across which to perform operation}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Reduce Prod Across Axis
}
