% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_reshape}
\alias{op_reshape}
\title{Reshape}
\usage{
op_reshape(x, shape, begin_axis = NULL, end_axis = NULL, name = name)
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{shape}{- list of ints representing tensor shape}

\item{begin_axis}{- shape replacement begins at this axis}

\item{end_axis}{- shape replacement ends at this axis (non-inclusive)}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Reshape
}
