% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_roipooling}
\alias{op_roipooling}
\title{Region of Interest Pooling}
\usage{
op_roipooling(conv_feature_map, rois, roi_output_shape, name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Region of Interest Pooling
}
