% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_round}
\alias{op_round}
\title{Element-wise Rounding}
\usage{
op_round(arg, name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Element-wise Rounding
}
