% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_sin}
\alias{op_sin}
\title{Element-wise Sine}
\usage{
op_sin(x, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Element-wise Sine
}
