% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_slice}
\alias{op_slice}
\title{Slice}
\usage{
op_slice(x, axis, begin_index, end_index, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{axis}{- axis across which to perform operation}

\item{begin_index}{- index where slicing starts}

\item{end_index}{- index where slicing ends}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Slice
}
