% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_softplus}
\alias{op_softplus}
\title{Softplus}
\usage{
op_softplus(x, steepness = 1, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Softplus operation. Computes the element-wise softplus of x:
}
\details{
softplus(x)=log(1+exp(x))softplus(x)=log⁡(1+exp⁡(x))

The optional steepness allows to make the knee sharper (steepness>1) or
softer, by computing softplus(x * steepness) / steepness. (For very large
steepness, this approaches a linear rectifier).

The output tensor has the same shape as x.
}
