% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_splice}
\alias{op_splice}
\title{Concatenate Across Axis}
\usage{
op_splice(inputs, ...)
}
\arguments{
\item{inputs}{- one or more input tensors}

\item{...}{named list of axis_names}
}
\description{
Concatenate Across Axis
}
