% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_stop_gradient}
\alias{op_stop_gradient}
\title{Stop Gradient}
\usage{
op_stop_gradient(input, name = "")
}
\arguments{
\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Stop Gradient
}
