% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_swap_axes}
\alias{op_swap_axes}
\title{Swap Axes}
\usage{
op_swap_axes(x, axis1 = 0, axis2 = 1, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Swap Axes
}
