% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_tanh}
\alias{op_tanh}
\title{Hyperbolic tan}
\usage{
op_tanh(x, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Hyperbolic tan
}
