% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_times}
\alias{op_times}
\title{Matrix Product}
\usage{
op_times(left, right, output_rank = 1, infer_input_rank_to_map = -1,
  name = "")
}
\arguments{
\item{left}{- left side tensor}

\item{right}{- right side tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Matrix Product
}
