% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_times_transpose}
\alias{op_times_transpose}
\title{One Element Times Another Transposed Element}
\usage{
op_times_transpose(left, right, name = "")
}
\arguments{
\item{left}{- left side tensor}

\item{right}{- right side tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
One Element Times Another Transposed Element
}
