% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_to_sequence}
\alias{op_to_sequence}
\title{To Sequence}
\usage{
op_to_sequence(x, sequence_lengths = NULL,
  sequence_axis_name_prefix = "toSequence_", name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
To Sequence
}
