% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_to_sequence_like}
\alias{op_to_sequence_like}
\title{To Sequence Like}
\usage{
op_to_sequence_like(x, dynamic_axes_like, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
To Sequence Like
}
