% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_transpose}
\alias{op_transpose}
\title{Transpose}
\usage{
op_transpose(x, perm, name = "")
}
\arguments{
\item{x}{- matrix or CNTK Function that outputs a tensor}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Transpose
}
