% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{op_unpooling}
\alias{op_unpooling}
\title{Unpooling}
\usage{
op_unpooling(operand, pooling_input, unpooling_window_shape, strides = c(1),
  auto_padding = c(FALSE), name = "")
}
\arguments{
\item{operand}{}

\item{strides}{(int or tuple of ints, defaults to 1) – stride of the
operation. Use a list of ints to specify a per-axis value.}

\item{name}{(str) - the name of the Function instance in the network}
}
\description{
Unpooling
}
