% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Parameter}
\alias{Parameter}
\title{Parameter}
\usage{
Parameter(value = NULL, shape = NULL, dtype = "float32", device = NULL,
  name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{dtype}{- data type to be used ("float32", "float64", or "auto")}

\item{device}{- instance of DeviceDescriptor}

\item{name}{}
}
\description{
A trainable parameter. It can be a scalar, vector, matrix, or tensor of
floating point numbers that can be modified by a training procedure.
}
\details{
****** Properties: ******

value
}
