% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-sequence.R
\name{PastValueWindow}
\alias{PastValueWindow}
\title{Past Value Window}
\usage{
PastValueWindow(window_size, axis, go_backwards = FALSE, name = "")
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Past Value Window
}
