% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{printer_log}
\alias{printer_log}
\title{Printer - Log}
\usage{
printer_log(printer, message)
}
\arguments{
\item{printer}{Printer instance on which to perform the operation}

\item{message}{}
}
\description{
Printer - Log
}
