% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{printer_on_training_update_end}
\alias{printer_on_training_update_end}
\title{Printer - On Training Update End}
\usage{
printer_on_training_update_end(printer)
}
\arguments{
\item{printer}{}
}
\description{
Printer - On Training Update End
}
