% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{printer_on_write_distributed_sync_update}
\alias{printer_on_write_distributed_sync_update}
\title{Printer - On Write Distributed Sync Update}
\usage{
printer_on_write_distributed_sync_update(printer, samples, updates,
  aggregate_metric)
}
\arguments{
\item{aggregate_metric}{}
}
\description{
Printer - On Write Distributed Sync Update
}
