% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{printer_on_write_test_summary}
\alias{printer_on_write_test_summary}
\title{any writer}
\usage{
printer_on_write_test_summary(printer, samples, updates, summaries,
  aggregate_metric, elapsed_milliseconds)
}
\description{
any writer
}
