% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{printer_write}
\alias{printer_write}
\title{Printer - Write}
\usage{
printer_write(printer, key, value)
}
\arguments{
\item{printer}{Printer instance on which to perform the operation}

\item{value}{}
}
\description{
Printer - Write
}
