% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{ProgressPrinter}
\alias{ProgressPrinter}
\title{Create Progress Printer}
\usage{
ProgressPrinter(freq = NULL, first = 0, tag = "", log_to_file = NULL,
  rank = NULL, gen_heartbeat = FALSE, num_epochs = NULL,
  test_freq = NULL, test_first = 0, metric_is_pct = TRUE,
  distributed_freq = NULL, distributed_first = 0)
}
\arguments{
\item{distributed_first}{}
}
\description{
Create Progress Printer
}
