% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rand_bernoulli}
\alias{rand_bernoulli}
\title{Random Bernoulli Distribution}
\usage{
rand_bernoulli(shape, dtype = default_override_or(np$float32), mean = 0.5,
  seed = AUTO_SELECT_SEED, name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{dtype}{- data type to be used ("float32", "float64", or "auto")}

\item{mean}{- success probability}

\item{seed}{- pseudo random number generator seed}

\item{name}{- name of the Function instance in the network}
}
\description{
Generates samples from the Bernoulli distribution with success probability
mean.
}
