% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rand_bernoulli_like}
\alias{rand_bernoulli_like}
\title{Random Bernoulli Like}
\usage{
rand_bernoulli_like(x, mean = 0.5, seed = AUTO_SELECT_SEED, name = "")
}
\arguments{
\item{x}{- CNTK variable from which to copy the shape, dtype, and dynamic axes}

\item{mean}{- success probability}

\item{seed}{- pseudo random number generator seed}

\item{name}{- name of the Function instance in the network}
}
\description{
Generates samples from the Bernoulli distribution with success probability
mean.
}
