% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rand_gumbel}
\alias{rand_gumbel}
\title{Random Gumbel Distribution}
\usage{
rand_gumbel(shape, dtype = default_override_or(np$float32), loc = 0,
  scale = 1, seed = AUTO_SELECT_SEED, name = name)
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{dtype}{- data type to be used ("float32", "float64", or "auto")}

\item{scale}{- scale of the distribution}

\item{seed}{- pseudo random number generator seed}

\item{name}{- name of the Function instance in the network}
}
\description{
Random Gumbel Distribution
}
