% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rand_gumbel_like}
\alias{rand_gumbel_like}
\title{Random Gumbel Like Distribution}
\usage{
rand_gumbel_like(x, mean = 0, scale = 1, seed = AUTO_SELECT_SEED,
  name = "")
}
\arguments{
\item{x}{- CNTK variable from which to copy the shape, dtype, and dynamic axes}

\item{mean}{- success probability}

\item{scale}{- scale of the distribution}

\item{seed}{- pseudo random number generator seed}

\item{name}{- name of the Function instance in the network}
}
\description{
Random Gumbel Like Distribution
}
