% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rand_uniform}
\alias{rand_uniform}
\title{Random Uniform Distribution}
\usage{
rand_uniform(shape, dtype = default_override_or(np$float32), low = 0,
  high = 1, seed = AUTO_SELECT_SEED, name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{dtype}{- data type to be used ("float32", "float64", or "auto")}

\item{low}{- bottom of distribution range}

\item{high}{- top of distribution range}

\item{seed}{- pseudo random number generator seed}

\item{name}{- name of the Function instance in the network}
}
\description{
Random Uniform Distribution
}
