% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rand_uniform_like}
\alias{rand_uniform_like}
\title{Random Uniform Like Distribution}
\usage{
rand_uniform_like(x, low = 0, high = 1, seed = AUTO_SELECT_SEED,
  name = "")
}
\arguments{
\item{x}{- CNTK variable from which to copy the shape, dtype, and dynamic axes}

\item{low}{- bottom of distribution range}

\item{high}{- top of distribution range}

\item{seed}{- pseudo random number generator seed}

\item{name}{- name of the Function instance in the network}
}
\description{
Random Uniform Like Distribution
}
