% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Record}
\alias{Record}
\title{Record}
\usage{
Record(...)
}
\arguments{
\item{...}{named arguments to turn into record numbers}
}
\description{
Easy construction of a record (=immutable singleton class) from keyword
arguments.
}
\details{
****** Associated Functions: ******

updated_record_with
}
