% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-sequence.R
\name{RecurrenceFrom}
\alias{RecurrenceFrom}
\title{Recurrence From}
\usage{
RecurrenceFrom(step_function, go_backwards, return_full_state, name = "")
}
\arguments{
\item{step_function}{}

\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Recurrence From
}
