% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{register_native_user_function}
\alias{register_native_user_function}
\title{Register Native UserFunction}
\usage{
register_native_user_function(op_id, module_name, factory_method_name)
}
\arguments{
\item{factory_method_name}{}
}
\description{
Registers a native user-defined Function that can be subsequently
instantiated using the ‘native_user_function’ method.
}
