% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{register_udf_deserialize_callback}
\alias{register_udf_deserialize_callback}
\title{Register UDF Deserialize Callback}
\usage{
register_udf_deserialize_callback(op_name, callback)
}
\arguments{
\item{callback}{}
}
\description{
Register a callback function to be invoked when deserializing a user-
defined function with the corresponding op name.
}
\details{
When loading a model, CNTK will try to automatically reconstruct any
(non-native) user-defined functions by invoking a static deserialize()
method of the corresponding UserFunction sub-class. This method allows to
override default UDF deserialization behavior by specifying a user- defined
function op name and the corresponding callback that should be invoked
instead of the deserialize method.
}
