% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{reset_learning_rate}
\alias{reset_learning_rate}
\title{Reset Learner Learning Rate}
\usage{
reset_learning_rate(learner, learning_rate)
}
\arguments{
\item{learning_rate}{}
}
\description{
Reset Learner Learning Rate
}
