% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{restore_mb_from_checkpoint}
\alias{restore_mb_from_checkpoint}
\title{restore_mb_from_checkpoint}
\usage{
restore_mb_from_checkpoint(mb_source, checkpoint)
}
\arguments{
\item{checkpoint}{}
}
\description{
Sets the state of the checkpoint.
}
