% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{restore_trainer_from_checkpoint}
\alias{restore_trainer_from_checkpoint}
\title{Restore Trainer From Checkpoint}
\usage{
restore_trainer_from_checkpoint(trainer, filename)
}
\arguments{
\item{filename}{}
}
\description{
Restore Trainer From Checkpoint
}
