% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging-debug.R
\name{save_as_legacy_model}
\alias{save_as_legacy_model}
\title{Save As Legacy Model}
\usage{
save_as_legacy_model(root_op, filename)
}
\arguments{
\item{filename}{}
}
\description{
Save As Legacy Model
}
