% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{save_trainer_checkpoint}
\alias{save_trainer_checkpoint}
\title{Save Trainer Checkpoint}
\usage{
save_trainer_checkpoint(trainer, filename)
}
\arguments{
\item{filename}{}
}
\description{
Save Trainer Checkpoint
}
