% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-sequence.R
\name{seq_broadcast_as}
\alias{seq_broadcast_as}
\title{Broadcast Sequence As}
\usage{
seq_broadcast_as(operand, broadcast_as_operand, name = "")
}
\arguments{
\item{name}{}
}
\description{
Broadcast Sequence As
}
