% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-sequence.R
\name{seq_future_value}
\alias{seq_future_value}
\title{Sequence Future Value}
\usage{
seq_future_value(x, initial_state = NULL, time_step = 1, name = "")
}
\arguments{
\item{name}{}

\item{init}{(scalar or matrix or initializer, defaults to
init_glorot_uniform()) – initial value of weights Wial_state}
}
\description{
Sequence Future Value
}
