% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-sequence.R
\name{seq_input_variable}
\alias{seq_input_variable}
\title{Sequence Network Input Variable}
\usage{
seq_input_variable(shape, dtype = "float32", needs_gradient = FALSE,
  is_sparse = FALSE, sequence_axis = CNTKAxis$default_dynamic_axis(),
  name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{dtype}{- data type to be used ("float32", "float64", or "auto")}

\item{name}{}
}
\description{
Sequence Network Input Variable
}
