% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-sequence.R
\name{seq_scatter}
\alias{seq_scatter}
\title{Sequence Scatter}
\usage{
seq_scatter(seq, condition, new_sequence_axis_typeinfo = NULL, name = "")
}
\arguments{
\item{name}{}
}
\description{
Sequence Scatter
}
