% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{sequence_to_cntk_text_format}
\alias{sequence_to_cntk_text_format}
\title{Base64ImageDeserializer}
\usage{
sequence_to_cntk_text_format(seq_inx, alias_tensor_map)
}
\description{
****** Attributes: ******
}
\details{
filename
}
