% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-higher_order_layers.R
\name{SequentialClique}
\alias{SequentialClique}
\title{Sequential Clique}
\usage{
SequentialClique(..., name = "")
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Sequential Clique
}
