% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{sess_minibatch_size_schedule}
\alias{sess_minibatch_size_schedule}
\title{Create a Minibatch Size Schedule}
\usage{
sess_minibatch_size_schedule(schedule, epoch_size = 1)
}
\arguments{
\item{epoch_size}{}
}
\description{
Create a Minibatch Size Schedule
}
