% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{sess_training_session}
\alias{sess_training_session}
\title{Create Training Session Object}
\usage{
sess_training_session(trainer, mb_source, mb_size, model_inputs_to_streams,
  progress_frequency = NULL, max_samples = NULL, checkpoint_config = NULL,
  cv_config = NULL, test_config = NULL)
}
\arguments{
\item{test_config}{}
}
\description{
Create Training Session Object
}
