% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-blocks.R
\name{Stabilizer}
\alias{Stabilizer}
\title{Stabilizer}
\usage{
Stabilizer(steepness = 4, enable_self_stabilization = TRUE, name = "")
}
\description{
Layer factory function to create a Droppo self-stabilizer. It multiplies its
input with a scalar that is learned.  This takes enable_self_stabilization
as a flag that allows to disable itself. Useful if this is a global default.
}
