% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging-profiler.R
\name{start_profiler}
\alias{start_profiler}
\title{Start Profiler}
\usage{
start_profiler(dir = "profiler", sync_gpu = TRUE, reserve_mem = 33554432)
}
\arguments{
\item{reserve_mem}{}
}
\description{
Start Profiler
}
