% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{StreamInformation}
\alias{StreamInformation}
\title{StreamInformation}
\usage{
StreamInformation(name, stream_id, storage_format, shape)
}
\description{
Stream information container that is used to describe streams when
implementing custom minibatch source through UserMinibatchSource.
}
