% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{tensorboard_close}
\alias{tensorboard_close}
\title{Close Tensorboard}
\usage{
tensorboard_close(tensorboard_writer)
}
\arguments{
\item{tensorboard_writer}{}
}
\description{
Close Tensorboard
}
