% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{tensorboard_write_value}
\alias{tensorboard_write_value}
\title{TensorBoard Write Value}
\usage{
tensorboard_write_value(tensorboard_writer, name, value, step)
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}

\item{step}{}
}
\description{
TensorBoard Write Value
}
