% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{TensorBoardProgressWriter}
\alias{TensorBoardProgressWriter}
\title{New TensorBoard Progress Writer}
\usage{
TensorBoardProgressWriter(freq = NULL, log_dir = ".", rank = NULL,
  model = NULL)
}
\arguments{
\item{model}{}
}
\description{
New TensorBoard Progress Writer
}
