% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{TestConfig}
\alias{TestConfig}
\title{New Test Configuration}
\usage{
TestConfig(minibatch_source = NULL, minibatch_size = 32,
  model_inputs_to_streams = NULL, criterion = NULL, source = NULL,
  mb_size = NULL)
}
\arguments{
\item{mb_size}{}
}
\description{
New Test Configuration
}
