% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{train_on_session}
\alias{train_on_session}
\title{Train On TrainingSession}
\usage{
train_on_session(sess, device = NULL)
}
\arguments{
\item{sess}{- session instance on which to perform the operation}

\item{device}{- instance of DeviceDescriptor}
}
\description{
Train On TrainingSession
}
