% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{Trainer}
\alias{Trainer}
\title{Trainer}
\usage{
Trainer(model, criterion, parameter_learners, progress_writers = NULL)
}
\arguments{
\item{model}{- root node of the Function to train}

\item{criterion}{(list of Function or Variable) - Function with one or two
outputs, representing loss and, if given, evaluation metric (in this order)}

\item{parameter_learners}{(list) – list of learners}

\item{progress_writers}{(progress writer or list of them) – optionally, list
of progress writers to automatically track training progress.}
}
\description{
The instance of the class should be created by using training_session()
function.
}
\details{
A training session trains a model using the specified trainer and configs.
Different aspects of training such as data sources, checkpointing, cross
validation, progress printing can be configured using the corresponding
config classes.

****** Properties: ******

evaluation_function - The evaluation function that the trainer is using.

loss_function - The loss function that the trainer is using.

model - The model that the trainer is training.

parameter_learners - The parameter learners that the trainer is using.

previous_minibatch_evaluation_average - The average evaluation criterion
value per sample for the last minibatch trained

previous_minibatch_loss_average - The average training loss per sample for
the last minibatch trained

previous_minibatch_sample_count - The number of samples in the last
minibatch trained with

total_number_of_samples_seen - The number of samples seen globally between
all workers from the beginning of training.

****** Associated Functions: ******

restore_trainer_from_checkpoint

save_trainer_checkpoint

summarize_training_progress

summarize_test_progress

test_minibatch

train_minibatch
}
