% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{TrainingSession}
\alias{TrainingSession}
\title{New Training Configuration}
\usage{
TrainingSession(trainer, mb_source, mb_size, model_inputs_to_streams,
  max_samples, progress_frequency, checkpoint_config, cv_config, test_config)
}
\arguments{
\item{trainer}{trainer}

\item{mb_source}{minibatch source used for training}

\item{mb_size}{minibatch size schedule for training}

\item{model_inputs_to_streams}{(dict): mapping between input variables and input streams}

\item{max_samples}{(int): maximum number of samples used for training}

\item{progress_frequency}{(int): frequency in samples for aggregated progress printing}

\item{checkpoint_config}{(:class:`CheckpointConfig`): checkpoint configuration}

\item{cv_config}{(:class:`CrossValidationConfig`): cross validation configuration}

\item{test_config}{(:class:`TestConfig`): test configuration}
}
\description{
The instance of the class should be created by using training_session()
function.  A training session trains a model using the specified trainer and
configs. Different aspects of training such as data sources, checkpointing,
cross validation, progress printing can be configured using the
corresponding config classes.
}
\details{
****** Associated Functions: ******

on_train_cross_validation_end

train_on_session
}
