% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{TrainingSummaryProgressCallback}
\alias{TrainingSummaryProgressCallback}
\title{Training Summary Progress Callback}
\usage{
TrainingSummaryProgressCallback(epoch_size, callback)
}
\arguments{
\item{callback}{}
}
\description{
Training Summary Progress Callback
}
