% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-transforms.R
\name{transform_color}
\alias{transform_color}
\title{Color Transform}
\usage{
transform_color(brightness_radius = 0, contrast_radius = 0,
  saturation_radius = 0)
}
\arguments{
\item{saturation_radius}{}
}
\description{
Color Transform
}
