% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-transforms.R
\name{transform_crop}
\alias{transform_crop}
\title{Crop Transform}
\usage{
transform_crop(crop_type = "center", crop_size = 0, side_ratio = 0,
  area_ratio = 0, aspect_ratio = 1, jitter_type = "none")
}
\arguments{
\item{crop_type}{}

\item{jitter_type}{}
}
\description{
Crop Transform
}
