% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-transforms.R
\name{transform_scale}
\alias{transform_scale}
\title{Scale Transform}
\usage{
transform_scale(width, height, channels, interpolations, scale_mode, pad_value)
}
\arguments{
\item{width}{}

\item{pad_value}{}
}
\description{
Scale Transform
}
