% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{try_set_default_device}
\alias{try_set_default_device}
\title{Try Set Default Device}
\usage{
try_set_default_device(new_default_device, acquire_device_lock = FALSE)
}
\arguments{
\item{acquire_device_lock}{}
}
\description{
Try Set Default Device
}
