% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-sequence.R
\name{UnfoldFrom}
\alias{UnfoldFrom}
\title{Unfold From}
\usage{
UnfoldFrom(generator_function, until_predicate = NULL, length_increase = 1,
  name = "")
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Unfold From
}
