% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-blocks.R
\name{UntestedBranchError}
\alias{UntestedBranchError}
\title{Untested Branch Error}
\usage{
UntestedBranchError(name)
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Untested Branch Error
}
