% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{updated_record_with}
\alias{updated_record_with}
\title{Updated Record With}
\usage{
updated_record_with(record, ...)
}
\arguments{
\item{record}{- the Record instance to be updated}

\item{...}{named arguments to turn into record numbers}
}
\description{
Create a new Record from an existing one with members modified or added.
}
