% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{user_function}
\alias{user_function}
\title{User Defined Function}
\usage{
user_function(user_func)
}
\arguments{
\item{user_func}{User Function}
}
\description{
Wraps the passed Function to create a composite representing the composite
Function graph rooted at the passed root Function.
}
