% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{userfunc_clone}
\alias{userfunc_clone}
\title{Clone UserFunction}
\usage{
userfunc_clone(func, cloned_inputs)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}

\item{cloned_inputs}{}
}
\description{
Creates a clone of this user-defined function.
}
\details{
It assumes that the constructor signature of the user’s implementation of
the user function takes the inputs as individual arguments followed by the
operator name. If the signature is different, then this method needs to be
overriden.
}
