% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{userfunc_deserialize}
\alias{userfunc_deserialize}
\title{Deserialize UserFunction}
\usage{
userfunc_deserialize(inputs, name, state)
}
\arguments{
\item{state}{}
}
\description{
A stub deserialize method for illustration purposes. User-defined functions
need to provide their own implementation in order for CNTK to be able to
reconstruct them when loading a model.
}
