% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{userfunc_infer_outputs}
\alias{userfunc_infer_outputs}
\title{Infer Outputs of UserFunction}
\usage{
userfunc_infer_outputs(func)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}
}
\description{
Returns a list of all output variables this user-defined function outputs.
Output variables are created by output_variable().
}
