% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{userfunc_serialize}
\alias{userfunc_serialize}
\title{Serialize UserFunction}
\usage{
userfunc_serialize(func)
}
\arguments{
\item{func}{- The CNTK `Function` instance on which to apply the operation}
}
\description{
Generates a dictionary that captures the state of this user-defined
function.  This method must be overridden, if a user function has any state
that needs to be preserved in the model dictionary.
}
