% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-function.R
\name{UserFunction}
\alias{UserFunction}
\title{User Function}
\usage{
UserFunction(inputs, as_matrix = TRUE, name = "")
}
\arguments{
\item{as_matrix}{- whether to return as an R matrix. Defualt TRUE. Otherwise
returns as Python CNTK value which avoids costly conversions}

\item{name}{}
}
\description{
Base class of all user extension functions.
}
\details{
If it has only one output, one can invoke Variable methods on it, which it
will relay to its only output.

See ?Function for more info

****** Special Functions: ******

userfunc_clone

userfunc_deserialize

userfunc_infer_outputs

userfunc_serialize
}
