% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{UserMinibatchSource}
\alias{UserMinibatchSource}
\title{UserMinibatchSource}
\usage{
UserMinibatchSource()
}
\description{
Base class of all user minibatch sources.
}
\details{
****** Associated Functions: ******

get_minibatch_checkpoint_state

next_minibatch

restore_mb_from_checkpoint

mb_stream_info

mb_stream_infos
}
