% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{value_as_matrix}
\alias{value_as_matrix}
\title{Value As Matrix}
\usage{
value_as_matrix(value, dtype = "auto")
}
\arguments{
\item{value}{Value instance to be converted}

\item{dtype}{data type to be used ("float32", "float64", or "auto")}
}
\description{
Converts a Value object to a sequence of matrices (if dense) or CSR arrays
(if sparse).
}
