% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{value_as_sequences}
\alias{value_as_sequences}
\title{Get Value As Sequence of Matrices}
\usage{
value_as_sequences(value, variable = NULL)
}
\arguments{
\item{value}{the Value instance}

\item{variable}{the Variable}
}
\description{
Convert a Value to a sequence of NumPy arrays that have their masked entries removed.
}
