% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{value_create}
\alias{value_create}
\title{Create a Value Object}
\usage{
value_create(var, data, seq_starts = NULL, device = NULL,
  read_only = FALSE)
}
\arguments{
\item{var}{(Variable) variable into which data is passed}

\item{data}{matrix denoting the full minibatch or one parameter or constant}

\item{seq_starts}{(list of bools or NULL) if None, every sequence is treated
as a new sequence. Otherwise, it is interpreted as a list of Booleans that
tell whether a sequence is a new sequence (True) or a continuation of the
sequence in the same slot of the previous minibatch (False)}

\item{device}{instance of DeviceDescriptor}

\item{read_only}{(bool) whether the data can be modified}
}
\description{
Creates a Value object
}
