% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{value_one_hot}
\alias{value_one_hot}
\title{Value as One Hot}
\usage{
value_one_hot(batch, num_classes, dtype = "auto", device = NULL)
}
\arguments{
\item{batch}{(list of list of int) batch input data of indices}

\item{num_classes}{(int or list)  number of classes or shape of each sample
whose trailing axis is one_hot}

\item{dtype}{data type to be used ("float32", "float64", or "auto")}

\item{device}{instance of DeviceDescriptor}
}
\description{
Converts batch into a Value object of dtype such that the integer data in
batch is interpreted as the indices representing one-hot vectors.
}
