% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Variable}
\alias{Variable}
\title{Variable}
\usage{
Variable(shape = NULL, dtype = "auto", needs_gradient = FALSE,
  is_sparse = FALSE, dynamic_axes = rep(c(get_default_batch_axis()), 2),
  name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{dtype}{- data type to be used ("float32", "float64", or "auto")}

\item{name}{}
}
\description{
Denotes a symbolic entity corresponding to the inputs and outputs of a
Function.
}
