% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{VariableMixin}
\alias{VariableMixin}
\title{Variable Mixin}
\usage{
VariableMixin()
}
\description{
Standard properties for Variable and its derived classes Parameter and
Constant.
}
\details{
****** Properties: ******

dtype

dynamic_axes

is_constant

is_input

is_output

is_parameter

is_placeholder

is_sparse

name

needs_gradient

owner

shape

uid
}
