% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize Network Architecture}
\usage{
visualize_network(model)
}
\arguments{
\item{model}{CNTK network}
}
\value{
PNG plot
}
\description{
Visualize the network architecture as a dataflow graph
Describes the operations and how they interact with each other
}
