% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-distributed.R
\name{WorkerDescriptor}
\alias{WorkerDescriptor}
\title{New Distributed Worker Descriptor}
\usage{
WorkerDescriptor()
}
\description{
Distributed worker descriptor, returned by Communicator instance.
}
\details{
****** Properties: ******

global_rank

host_id
}
