% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-layers.R
\name{BatchNormalization}
\alias{BatchNormalization}
\title{BatchNormalization}
\usage{
BatchNormalization(map_rank = NULL, init_scale = 1,
  normalization_time_constant = 5000, blend_time_constant = 0,
  epsilon = 1e-05, use_cntk_engine = FALSE, name = "")
}
\arguments{
\item{name}{string for the name of the function instance in the network}
}
\description{
Layer factory function to create an average-pooling layer.
}
\details{
Like Convolution(), AveragePooling() processes items arranged on an
N-dimensional grid, such as an image. Typically, each item is a vector. For
each item, average-pooling computes the element-wise mean over a window
(“receptive field”) of items surrounding the item’s position on the grid.

The size (spatial extent) of the receptive field is given by filter_shape.
E.g. for 2D pooling, filter_shape should be a tuple of two integers, such as
(5,5).
}
