% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{CheckpointConfig}
\alias{CheckpointConfig}
\title{New Checkpoint Config}
\usage{
CheckpointConfig(filename, frequency = NULL, restore = TRUE,
  preserve_all = FALSE)
}
\arguments{
\item{preserve_all}{}
}
\description{
New Checkpoint Config
}
