% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-layers.R
\name{Convolution2D}
\alias{Convolution2D}
\title{Convoluion2D}
\usage{
Convolution2D(filter_shape, num_filters = NULL,
  activation = activation_identity, init = init_glorot_uniform(),
  pad = FALSE, strides = 1, bias = TRUE, init_bias = 0,
  reduction_rank = 1, name = "")
}
\description{
Layer factory function to create a 2D convolution layer with optional
non-linearity. Same as Convolution() except that filter_shape is verified to
be 2-dimensional. See Convolution() for extensive documentation.
}
