% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{CTFDeserializer}
\alias{CTFDeserializer}
\title{CTFDeserializer}
\usage{
CTFDeserializer(filename, streams)
}
\arguments{
\item{filename}{A string containing the path to the data location}

\item{streams}{A python dictionary-like object that contains a mapping from stream names to StreamDef objects. Each StreamDef object configures an input stream.}
}
\description{
Configures the CNTK text-format reader that reads text-based files.
}
\details{
Form: `[Sequence_Id] (Sample)+` where `Sample=|Input_Name (Value )*`
}
\references{
See also \url{https://www.cntk.ai/pythondocs/cntk.io.html?highlight=ctfdeserializer#cntk.io.CTFDeserializer}
}
