% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_test_minibatch}
\alias{eval_test_minibatch}
\title{Test Evaluator Minibatch}
\usage{
eval_test_minibatch(evaluator, arguments, device = NULL)
}
\arguments{
\item{evaluator}{}

\item{device}{- instance of DeviceDescriptor}
}
\description{
Test Evaluator Minibatch
}
