% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{HTKFeatureDeserializer}
\alias{HTKFeatureDeserializer}
\title{HTKFeatureDeserializer}
\usage{
HTKFeatureDeserializer(streams)
}
\description{
Configures the HTK feature reader that reads speech data from scp files.
}
