% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{HTKMLFDeserializer}
\alias{HTKMLFDeserializer}
\title{Base64ImageDeserializer}
\usage{
HTKMLFDeserializer(label_mapping_file, streams, phoneBoundaries = FALSE)
}
\arguments{
\item{phoneBoundaries}{}
}
\description{
Configures an HTK label reader that reads speech HTK format MLF (Master
Label File)
}
