% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ImageDeserializer}
\alias{ImageDeserializer}
\title{ImageDeserializer}
\usage{
ImageDeserializer(filename, streams)
}
\arguments{
\item{streams}{}
}
\description{
Configures the image reader that reads images and corresponding labels from
a file.
}
\details{
Form: `<full path to image> <tab> <numerical label (0-based class id)>` or
'sequenceId <tab> path <tab> label`
}
