% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{init_glorot_normal}
\alias{init_glorot_normal}
\title{Glorot Normal Initializer}
\usage{
init_glorot_normal(scale = 1, output_rank = 2147483647,
  filter_rank = 2147483647, seed = NULL)
}
\arguments{
\item{seed}{}
}
\description{
Glorot Normal Initializer
}
