% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-layers.R
\name{Label}
\alias{Label}
\title{Label}
\usage{
Label(name)
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}
}
\description{
Layer factory function to create a dummy layer with a given name. This can be
used to access an intermediate value flowing through computation.
}
