% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-layers.R
\name{LayerNormalization}
\alias{LayerNormalization}
\title{LayerNormalization}
\usage{
LayerNormalization(initial_scale = 1, initial_bias = 0, epsilon = 1e-05,
  name = "")
}
\arguments{
\item{name}{string for the name of the function instance in the network}
}
\description{
Layer factory function to create a function that implements layer
normalization.
}
\details{
Layer normalization applies this formula to every input element
(element-wise): y = (x - mean(x)) / (stddev(x) + epsilon) * scale + bias
where scale and bias are learned scalar parameters.
}
