% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{learner_momentum_sgd}
\alias{learner_momentum_sgd}
\title{Creates a Momentum SGD learner instance to learn the parameters.}
\usage{
learner_momentum_sgd(parameters, lr, momentum,
  unit_gain = cntk$default_unit_gain_value(), l1_regularization_weight = 0,
  l2_regularization_weight = 0, gaussian_noise_injection_std_dev = 0,
  gradient_clipping_threshold_per_sample = np$inf,
  gradient_clipping_with_truncation = TRUE, use_mean_gradient = FALSE)
}
\arguments{
\item{parameters}{list of network parameters to tune.}

\item{lr}{output of \code{learning_rate_schedule}}

\item{momentum}{output of \code{momentum_schdule} or \code{momentum_as_time_constant_schedule}}

\item{unit_gain}{logical whether to interpret momentum as a unit-gain filter}

\item{l1_regularization_weight}{double of l1 regularization}

\item{l2_regularization_weight}{double of l2 regularization}

\item{gaussian_noise_injection_std_dev}{double of noise injection}

\item{gradient_clipping_threshold_per_sample}{double of gradient clipping threshold per sample}

\item{gradient_clipping_with_truncation}{logical for gradient clipping with truncation}

\item{use_mean_gradient}{logical use averaged gradient as input to learner.}
}
\description{
Creates a Momentum SGD learner instance to learn the parameters.
}
\references{
\url{https://www.cntk.ai/pythondocs/cntk.learners.html#cntk.learners.momentum_sgd}
}
