% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers-blocks.R
\name{LSTM}
\alias{LSTM}
\title{LSTM}
\usage{
LSTM(shape, cell_shape = NULL, activation = op_tanh,
  use_peepholes = FALSE, init = init_glorot_uniform(), init_bias = 0,
  enable_self_stabilization = FALSE, name = "")
}
\arguments{
\item{shape}{- list of ints representing tensor shape}

\item{activation}{(Function) - optional activation Function}

\item{init}{(scalar or matrix or initializer, defaults to init_glorot_uniform()) – initial value of weights W}

\item{name}{string (optional) the name of the Function instance in the network}

\item{init}{(scalar or matrix or initializer, defaults to init_glorot_uniform()) – initial value of weights W_bias}
}
\description{
Layer factory function to create an LSTM block for use inside a recurrence.
The LSTM block implements one step of the recurrence and is stateless. It
accepts the previous state as its first two arguments, and outputs its new
state as a two-valued tuple (h,c).
}
