% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mb_stream_info}
\alias{mb_stream_info}
\title{mb_stream_info}
\usage{
mb_stream_info(mb_source, name)
}
\arguments{
\item{name}{}
}
\description{
Gets the description of the stream with given name. Throws an exception if
there are none or multiple streams with this same name.
}
