% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{next_minibatch}
\alias{next_minibatch}
\title{next_minibatch}
\usage{
next_minibatch(minibatch_source, minibatch_size_in_samples, input_map = NULL,
  device = NULL, num_data_partitions = NULL, partition_index = NULL)
}
\arguments{
\item{partition_index}{}
}
\description{
next_minibatch
}
