% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{printer_on_write_test_summary}
\alias{printer_on_write_test_summary}
\title{Printer - On Write Test Summary}
\usage{
printer_on_write_test_summary(printer, samples, updates, summaries,
  aggregate_metric, elapsed_milliseconds)
}
\arguments{
\item{elapsed_milliseconds}{}
}
\description{
Printer - On Write Test Summary
}
