% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging-progress_print.R
\name{printer_on_write_training_update}
\alias{printer_on_write_training_update}
\title{any writer}
\usage{
printer_on_write_training_update(printer, samples, updates, aggregate_loss,
  aggregate_metric)
}
\description{
any writer
}
