% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{restore_mb_from_checkpoint}
\alias{restore_mb_from_checkpoint}
\title{Restore Minibatch From Checkpoint}
\usage{
restore_mb_from_checkpoint(mb_source, checkpoint)
}
\arguments{
\item{mb_source}{minibatch source on which to perform the operation}

\item{checkpoint}{(dict) checkpoint to restore from}
}
\description{
Restores the MinibatchSource state from the specified checkpoint.
}
