% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{StreamConfiguration}
\alias{StreamConfiguration}
\title{StreamConfiguration}
\usage{
StreamConfiguration(name, dim, is_sparse = FALSE, stream_alias = "",
  defines_mb_size = FALSE)
}
\description{
Configuration of a stream in a text format reader.
}
