% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{StreamConfiguration}
\alias{StreamConfiguration}
\title{Stream Configuration}
\usage{
StreamConfiguration(name, dim, is_sparse = FALSE, stream_alias = "",
  defines_mb_size = FALSE)
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the network}

\item{defines_mb_size}{}
}
\description{
Configuration of a stream in a text format reader.
}
