% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{StreamDef}
\alias{StreamDef}
\title{StreamDef}
\usage{
StreamDef(field = NULL, shape = NULL, is_sparse = FALSE,
  transforms = NULL, context = NULL, scp = NULL, mlf = NULL,
  broadcast = NULL, defines_mb_size = FALSE)
}
\description{
Configuration of a stream for use with the builtin Deserializers. The
meanings of some configuration keys have a mild dependency on the exact
deserializer, and certain keys are meaningless for certain deserializers.
}
