% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{StreamInformation}
\alias{StreamInformation}
\title{StreamInformation}
\usage{
StreamInformation(name, stream_id, storage_format, dtype, shape)
}
\arguments{
\item{name}{string (optional) the name of the Function instance in the
network}

\item{stream_id}{(int) unique ID of the stream}

\item{storage_format}{"dense" or "sparse"}

\item{dtype}{data type to be used ("float32", "float64", or "auto")}

\item{shape}{list of ints representing tensor shape}
}
\description{
Stream information container that is used to describe streams when
implementing custom minibatch source through UserMinibatchSource.
}
