% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R, R/train.R
\name{summarize_test_progress}
\alias{summarize_test_progress}
\alias{summarize_test_progress}
\title{Summarize Evaluator Test Progress}
\usage{
summarize_test_progress(trainer)

summarize_test_progress(trainer)
}
\arguments{
\item{trainer}{the Trainer instance on which to perform the operation}

\item{evaluator}{the Evaluator instance from which to get progress}
}
\description{
Summarize Evaluator Test Progress

Summarize Test Progress
}
