% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{summarize_training_progress}
\alias{summarize_training_progress}
\title{Summarize Training Progress}
\usage{
summarize_training_progress(trainer)
}
\arguments{
\item{trainer}{}
}
\description{
Summarize Training Progress
}
