% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{test_minibatch}
\alias{test_minibatch}
\title{Test Minibatch}
\usage{
test_minibatch(trainer, arguments, device = NULL)
}
\arguments{
\item{device}{- instance of DeviceDescriptor}
}
\description{
Test Minibatch
}
