% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_minibatch}
\alias{train_minibatch}
\title{Train Minibatch}
\usage{
train_minibatch(trainer, data, outputs = NULL, device = NULL)
}
\arguments{
\item{trainer}{the Trainer instance on which to perform the operation}

\item{device}{- instance of DeviceDescriptor}
}
\description{
Train Minibatch
}
