% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train-training_session.R
\name{TrainingSession}
\alias{TrainingSession}
\title{New Training Configuration}
\usage{
TrainingSession(trainer, mb_source, mb_size, model_inputs_to_streams,
  max_samples, progress_frequency, checkpoint_config, cv_config, test_config)
}
\arguments{
\item{test_config}{}
}
\description{
The instance of the class should be created by using training_session()
function.  A training session trains a model using the specified trainer and
configs. Different aspects of training such as data sources, checkpointing,
cross validation, progress printing can be configured using the
corresponding config classes.
}
\details{
****** Associated Functions: ******

on_train_cross_validation_end

train_on_session
}
