% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{UserLearner}
\alias{UserLearner}
\title{UserLearner}
\usage{
UserLearner(parameters, lr_schedule, as_matrix = TRUE)
}
\arguments{
\item{as_matrix}{}
}
\description{
Base class of all user-defined learners. To implement your own learning
algorithm, derive from this class and override the \code{update()}.
}
\details{
Certain optimizers (such as AdaGrad) require additional storage. This can be
allocated and initialized during construction.
}
\seealso{
\code{\link{update_user_learner}}
}
