% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{usermb_next_minibatch}
\alias{usermb_next_minibatch}
\title{usermb_next_minibatch}
\usage{
usermb_next_minibatch(usermb_source, num_samples, num_workers, worker_rank,
  device = NULL)
}
\arguments{
\item{device}{}
}
\description{
Function to be implemented by the user
}
